# Specialize isinf(::IEEEFloat)