@oscardssmith 2023-01-04
I'm not fully sure why this is faster. Any assembly language people want to chime in? Looks great though.