@mikmoore 2023-01-04
The relevant x86-64 assembly from `code_native` on `Float64` input:
```
## original
        vsubsd  %xmm0, %xmm0, %xmm1
        vxorpd  %xmm2, %xmm2, %xmm2
        vcmpordpd       %xmm2, %xmm0, %xmm0
        vcmpneqpd       %xmm2, %xmm1, %xmm1
        vandpd  %xmm1, %xmm0, %xmm0
        vmovd   %xmm0, %eax
        andb    $1, %al
        retq

## == version
        movabsq $.LCPI0_0, %rax
        vandpd  (%rax), %xmm0, %xmm0
        movabsq $.LCPI0_1, %rax
        vucomisd        (%rax), %xmm0
        setae   %al
        retq

## === version
        movb    $63, %cl
        bzhiq   %rcx, %rax, %rax
        movabsq $9218868437227405312, %rcx      # imm = 0x7FF0000000000000
        cmpq    %rcx, %rax
        sete    %al
        retq
```

The original version just had 2 conditions to check (and didn't find a way to do it better) while the PR version only checks 1. The PR version only requires zeroing the top bit of the input and comparing to the literal `Inf`. All the instructions are pretty basic (I think), so I'd just go with the instruction count: 7 for the original and 5 for either the `==` or `===` versions here.

The inner loop of `any(isinf,::Vector{Float64})`:
```
## original
.LBB0_5:                                # %L31
                                        # =>This Inner Loop Header: Depth=1
        cmpq    %rsi, %rcx
        je      .LBB0_6
# %bb.7:                                # %L43
                                        #   in Loop: Header=BB0_5 Depth=1
        vmovsd  (%rdx,%rsi,8), %xmm1            # xmm1 = mem[0],zero
        incq    %rsi
        vucomisd        %xmm1, %xmm1
        jp      .LBB0_5
# %bb.8:                                # %L43
                                        #   in Loop: Header=BB0_5 Depth=1
        vsubsd  %xmm1, %xmm1, %xmm1
        vucomisd        %xmm0, %xmm1
        jne     .LBB0_9
        jnp     .LBB0_5

## === version
.LBB0_4:                                # %L26
                                        # =>This Inner Loop Header: Depth=1
        cmpq    %rdx, %rcx
        je      .LBB0_5
# %bb.6:                                # %L38
                                        #   in Loop: Header=BB0_4 Depth=1
        bzhiq   %r9, (%rsi,%rdx,8), %rdi
        incq    %rdx
        cmpq    %r8, %rdi
        jne     .LBB0_4
```

The `any` versions show the benefits more clearly, hoisting the `mov` instructions and leaving 10 instructions for the base version and 6 in the PR.