@oscardssmith 2023-01-04
In that case I think this is good to merge. I'm honestly pretty surprised that LLVM couldn't figure out the optimal solution here.