@stevengj 2023-01-04
On my machine, there is a faster method still:
```jl
# this PR
isinf2(x::Float64) = abs(x) === oftype(x, Inf)

# another option:
isinf3(x::Float64) = (reinterpret(UInt64, x) & 0x7fffffffffffffff) == 0x7ff0000000000000
```
gives (using the same test inputs as above):
```jl
julia> @btime isinf($s); @btime any(isinf,$t); @btime any(isinf,$v);
  3.395 ns (0 allocations: 0 bytes)
  4.886 ns (0 allocations: 0 bytes)
  104.723 ns (0 allocations: 0 bytes)

julia> @btime isinf2($s); @btime any(isinf2,$t); @btime any(isinf2,$v);
  3.056 ns (0 allocations: 0 bytes)
  3.806 ns (0 allocations: 0 bytes)
  75.546 ns (0 allocations: 0 bytes)

julia> @btime isinf3($s); @btime any(isinf3,$t); @btime any(isinf3,$v);
  3.169 ns (0 allocations: 0 bytes)
  3.728 ns (0 allocations: 0 bytes)
  55.550 ns (0 allocations: 0 bytes)
```